Worked outside of git for a long time, extensible files work very well, only one test won't pass (not including the tests that require directories).
Primarily worked on inode.c and inode.h, not sure what other files were changed (hopefully the rest is just the formatting changes from working from a base Pintos download again, but oh well).

If something breaks or doesn't work, it is likely in this commit, feel free to change anything other than the inode.c and inode.h, the rest were likely meant to only be changed temporarily anyway.

Also, there is still much cleanup that needs to be done. And I am horribly ashamed with the complexity and obscurity of this code, but most of the work was done while incredibly tired, please don't judge me too much.